requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
beanie==1.26.0
motor==3.4.0
pydantic[email]==2.7.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
phonenumbers==8.13.39
